Extend support for IeeeFlags and FloatingPointControl.
- Use LDC inline assembler for X86 and X86_64.
- Add support for PowerPC
- Add support for MIPS

The PowerPC implementation works. But note that the Phobos model has some shortcomings.
PowerPC has 5 different inexact exceptions. In the Phobos model only the exception summary
is returned. To enable exceptions, floating points exceptions have to be enabled in the MSR,
bits FE0 and FE1. This has to be done outside of the FloatingPointControl struct.

The MIPS implementation is untested.