Fix an std.math.poly() unit test.
The expression "6L * x" with 6 being a long literal (instead of an
intended real literal) and x of type double evaluates to a double.
poly() gets a 6.0L in the array and an x promoted to a real and so
obviously uses real precision.
On Linux x64, this results in a negligible difference.