std.variant: don't assert that target address != null if type size is 0
A unittest triggers in line 1918 when testing Variant!(void[0]) with LDC
(current merge-2.068 branch).

When taking the address of a truly empty type, such as T[0], LDC always
returns a null constant; nothing is allocated on the stack.
Disabling this assert in case the target type's size is 0 should be safe,
there's no private state which could be modified by the following
assignment (it may still have side effects though) - right? ;)