Fix testAliasedString() unit test helper
Many of the functions for which this wrapper is used return
string ranges. Comparing these for object identity does not
make a lot of sense.

This is actually an issue, because toUTF, which is used in quite
a few of the tested functions, contains a void-initialized buffer,
so the identity comparison is not guaranteed to succeed. The tests
just so happen to pass on current DMD regardless, but this is not
the case for an optimized LDC build.

This is a back-port of upstream commit 6b0be3b.